requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.135.1
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic-core==2.41.5
starlette==0.52.1
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.41.0
anthropic==0.84.0
python-dotenv==1.2.2
psycopg2-binary==2.9.11
python-multipart==0.0.22
sqlalchemy==2.0.48
